refactor: Moved all cli things to a cli module
